+++ packages/react-ui/storybook/stories/common/Dialog.stories.js
@@ -47,6 +47,12 @@ const options = {
         options: ['lg', 'md', 'sm', 'xl', 'xs', false]
       }
     }
+  },
+  parameters: {
+    design: {
+      type: 'figma',
+      url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=1534%3A28896'
+    }
   }
 };
 export default options;