+++ packages/react-ui/storybook/stories/organisms/InputFile.stories.js
@@ -47,7 +47,7 @@ const options = {
     },
     design: {
       type: 'figma',
-      url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=1159%3A24236'
+      url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=1534%3A36997'
     }
   }
 };